Added User model, routes, controller, jwt, and bcrypt

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+const { use } = require('../routes/goalRoutes')
 //@description GET goals
 //@route GET /api/goals
 //@access private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -19,7 +21,8 @@ const setGoal =  asyncHandler(async (req, res) => {
         }
 
         const goal = await Goal.create({
-            text: req.body.text
+            text: req.body.text,
+            user: req.user.id
         })
         res.status(201).json(goal)
     
@@ -35,7 +38,19 @@ const updGoal =  asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error ('goal not found')
     }
+    const user = await User.findById(req.user.id)
 
+    //check for user
+    if (!user) {
+      res.status(401)  
+        throw new Error ('user not found')
+    }
+
+        //logged in user should match the goal-user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)  
+        throw new Error ('unauthorized Request')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(goal, req.body, { new: true})
 
     res.status(200).json(`Updated goal ${updatedGoal.id}`)
@@ -49,6 +64,20 @@ const rmGoal =  asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error ('goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if (!user) {
+      res.status(401)  
+        throw new Error ('user not found')
+    }
+
+        //logged in user should match the goal-user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)  
+        throw new Error ('unauthorized Request')
     }
     await goal.deleteOne()
     res.status(200).json({'message' : `Deleted Goal ${goal.id}`})
